breidablik/requirements.txt
wheel
libconf
requests
python-dateutil
pylint
matplotlib
Pillow
RPi.GPIO
numpy
spidev
vito
